refactor: dashboardCache -> paymentCache
