+++ pages/payments/index.tsx
@@ -4,7 +4,7 @@ import * as SuperTokensConfig from '../../config/backendConfig'
 import Session from 'supertokens-node/recipe/session'
 import { GetServerSideProps } from 'next'
 import TableContainer from '../../components/TableContainer/TableContainer'
-import { ButtonDisplayData } from 'redis/dashboardCache'
+import { ButtonDisplayData } from 'redis/paymentCache'
 import Image from 'next/image'
 import Link from 'next/link'
 import XECIcon from 'assets/xec-logo.png'
